User resolver and entity, db connection
Set up the user resolver with one query, and wired up the entity to be
both the db schema via TypeORM and a GraphQL type. For now, it looks
like I can connect to the db. My firewall probably isn't set up right on
my current dev machine. It's something I'll work on when I get the time.